fix load from fixtures

+++ marathon_marathons/signals.py
@@ -9,26 +9,27 @@ from marathon_utils.google_maps import get_route_elevation
 
 @receiver(post_save, sender=models.MarathonRoute)
 def save_route_heights(sender, **kwargs):
-    route = kwargs["instance"]
-    created = kwargs["created"]
+    if not kwargs.get('raw', False):
+        route = kwargs["instance"]
+        created = kwargs["created"]
 
-    if route._map_changed:
-        heights = models.MarathonHeight.objects.filter(route=route)
-        for h in heights:
-            h.delete()
+        if route._map_changed:
+            heights = models.MarathonHeight.objects.filter(route=route)
+            for h in heights:
+                h.delete()
 
-    if route._map_changed or created:
-        parsed_map = ast.literal_eval(route.parsed_map)
+        if route._map_changed or created:
+            parsed_map = ast.literal_eval(route.parsed_map)
 
-        elevation = get_route_elevation(parsed_map["route"])
-        for p in elevation:
-            mh = models.MarathonHeight.objects.create(route=route,
-                                               lat=p["location"]["lat"],
-                                               long=p["location"]["lng"],
-                                               elevation=p["elevation"],
-                                               resolution=p["resolution"])
+            elevation = get_route_elevation(parsed_map["route"])
+            for p in elevation:
+                mh = models.MarathonHeight.objects.create(route=route,
+                                                   lat=p["location"]["lat"],
+                                                   long=p["location"]["lng"],
+                                                   elevation=p["elevation"],
+                                                   resolution=p["resolution"])
 
-        models.MarathonRoute.objects.filter(pk=route.pk).update(raw_elevation_data=str({"points": elevation}))
+            models.MarathonRoute.objects.filter(pk=route.pk).update(raw_elevation_data=str({"points": elevation}))
